Use CLI pipeline for codon alignment tests
